// lib/normal-challenge-media-policy.d.ts
export declare const NORMAL_MEDIA_LIMITS: Readonly<{
  imageCount: 3;
  imageBytes: number;
  imageMinWidth: 712;
  imageMinHeight: 430;
  imageMaxWidth: 4000;
  imageMaxHeight: 2416;
  videoCount: 1;
  videoBytes: number;
  videoMinWidth: 1280;
  videoMinHeight: 720;
  videoMaxDurationSeconds: 75;
}>;